ecommerce/requirements_ecommerce.txt
blinker==1.8.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
filelock==3.14.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
platformdirs==4.2.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
virtualenv==20.26.2
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
